pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "partcraft3d"
version = "0.1.0"
description = "Industrial-scale data pipeline for generating native 3D editing training pairs from part-level datasets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "trimesh>=4.0",
    "Pillow>=10.0",
    "scipy>=1.11",
    "openai>=1.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "plyfile>=1.0",
]

[project.optional-dependencies]
render = ["bpy>=3.6"]
dev = ["pytest>=7.0", "ruff>=0.1"]

[tool.setuptools.packages.find]
include = ["partcraft*"]
